requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
pydicom==3.0.1
requests==2.31.0
SimpleITK==2.3.1
torch==2.0.1
torchvision==0.15.2
monai==1.3.0
dashscope==1.13.0
openai==0.28.1
faiss-cpu==1.7.4
sentence-transformers==2.2.0
scipy==1.10.1
tqdm==4.65.0 